+++ DL/DLFunctions.cs
@@ -5,23 +5,23 @@ using System.Text;
 
 namespace dl.DL
 {
-    using UpdateWeightFunc = Action<ILayer, ILayer, Func<IEnumerable<Tuple<double, double>>, double>, ILearningData>;
+    using UpdateWeightFunc = Action<ILayer, ILayer, Func<IEnumerable<Tuple<float, float>>, float>, ILearningData>;
 
     public static class DLF
     {
         static Random r = new Random(DateTime.Now.Millisecond);
-        const double h = 1e-5;
-        const double hh = h * 2.0;
-        public static Func<double, double> Derivative(this Func<double, double> f)
+        const float h = 1e-5f;
+        const float hh = h * 2.0f;
+        public static Func<float, float> Derivative(this Func<float, float> f)
         {
             return x => (f(x + h) - f(x - h)) / hh;
         }
 
-        public static double GetRandom()
+        public static float GetRandom()
         {
             var x = r.NextDouble();
             var y = r.NextDouble();
-            return Math.Sqrt(-2.0 * Math.Log(x)) * Math.Cos(2.0 * Math.PI * y);
+            return (float)Math.Sqrt(-2.0 * Math.Log(x)) * (float)Math.Cos(2.0 * Math.PI * y);
         }
 
         public static IEnumerable<T> Shuffle<T>(IEnumerable<T> xs)
@@ -37,7 +37,7 @@ namespace dl.DL
         }
 
 
-        public static double CalcFunction(INode node)
+        public static float CalcFunction(INode node)
         {
             return
                 node.Links
@@ -45,25 +45,25 @@ namespace dl.DL
                     .Sum();
         }
 
-        public static double GetRandomWeight() => 0.01 * DLF.GetRandom();
+        public static float GetRandomWeight() => 0.01f * DLF.GetRandom();
 
-        public static double ErrorFunctionCrossEntropy(IEnumerable<Tuple<double, double>> result)
+        public static float ErrorFunctionCrossEntropy(IEnumerable<Tuple<float, float>> result)
         {
-            return -result.Sum(a => a.Item2 * Math.Log(Math.Max(a.Item1, 1e-7)));// + (1 - a.Item2) * Math.Log(1 - a.Item1));
+            return (float)-result.Sum(a => a.Item2 * Math.Log(Math.Max(a.Item1, 1e-7)));// + (1 - a.Item2) * Math.Log(1 - a.Item1));
         }
-        public static double ErrorFunction(IEnumerable<Tuple<double, double>> result)
+        public static float ErrorFunction(IEnumerable<Tuple<float, float>> result)
         {
-            return 0.5 * result.Sum(a => Math.Pow(a.Item1 - a.Item2, 2));
+            return 0.5f * result.Sum(a => (float)Math.Pow(a.Item1 - a.Item2, 2));
         }
 
-        public static IEnumerable<double> ReLU(IEnumerable<double> xs) => xs.Select(x => Math.Max(0, x)).ToArray();
+        public static IEnumerable<float> ReLU(IEnumerable<float> xs) => xs.Select(x => Math.Max(0, x)).ToArray();
 
-        public static IEnumerable<double> Sigmoid(IEnumerable<double> xs) => xs.Select(x => 1.0 / (1.0 + Math.Exp(-x))).ToArray();
+        public static IEnumerable<float> Sigmoid(IEnumerable<float> xs) => xs.Select(x => 1.0f / (1.0f + (float)Math.Exp(-x))).ToArray();
 
-        public static IEnumerable<double> SoftMax(IEnumerable<double> xs)
+        public static IEnumerable<float> SoftMax(IEnumerable<float> xs)
         {
             var max = xs.Max();
-            var exps = xs.Select(x => Math.Exp(x - max)).ToArray();
+            var exps = xs.Select(x => (float)Math.Exp(x - max)).ToArray();
             var sum = exps.Sum();
             return exps.Select(x => x / sum).ToArray();
 
@@ -72,7 +72,7 @@ namespace dl.DL
         /// 出力層の重み計算
         public static void UpdateWeightOfOutputLayer(ILayer layer
                                                     , ILayer _
-                                                    , Func<IEnumerable<Tuple<double, double>>, double> errorFunction
+                                                    , Func<IEnumerable<Tuple<float, float>>, float> errorFunction
                                                     , ILearningData data)
         {
             var ox = layer.Nodes.Select(x => x.GetValue()).ToArray();
@@ -83,14 +83,14 @@ namespace dl.DL
                 var node = item.x;
                 var index = item.index;
                 // 誤差関数の偏微分
-                Func<double, double> ef = (double x) =>
+                Func<float, float> ef = (float x) =>
                 {
                     var rx = ox.ToArray();
                     rx[index] = x;
                     return errorFunction(rx.Zip(data.Expected, Tuple.Create));
                 };
                 // 活性化関数の偏微分
-                Func<double, double> of = (double x) =>
+                Func<float, float> of = (float x) =>
                 {
                     var rx = ux.ToArray();
                     rx[index] = x;
@@ -115,7 +115,7 @@ namespace dl.DL
 
         public static void UpdateWeightOfSoftMax(ILayer layer
                                                     , ILayer _
-                                                    , Func<IEnumerable<Tuple<double, double>>, double> errorFunction
+                                                    , Func<IEnumerable<Tuple<float, float>>, float> errorFunction
                                                     , ILearningData data)
         {
             var ys = layer.Nodes.Select(x => x.GetValue()).ToArray();
@@ -136,7 +136,7 @@ namespace dl.DL
         }
 
         /// 出力層以外の重み計算
-        public static UpdateWeightFunc UpdateWeight(Func<double, bool> ignore = null, Action<INode, INodeLink, double> update = null)
+        public static UpdateWeightFunc UpdateWeight(Func<float, bool> ignore = null, Action<INode, INodeLink, float> update = null)
         {
             ignore = ignore ?? (_ => false);
             update = update ?? Update;
@@ -145,7 +145,7 @@ namespace dl.DL
                 var ux = layer.Nodes.Select(layer.CalcFunction).ToArray();
 
                 // 活性化関数の偏微分
-                Func<int, double> of = (int index) =>
+                Func<int, float> of = (int index) =>
                 {
                     var temp = ux[index];
                     ux[index] = temp + h;
@@ -183,7 +183,7 @@ namespace dl.DL
             };
         }
 
-        static void Update(INode node, INodeLink link, double delta)
+        static void Update(INode node, INodeLink link, float delta)
         {
             // 前の層の出力
             var o0 = link.InputNode.GetValue();
@@ -192,10 +192,10 @@ namespace dl.DL
             link.InputNode.Delta += delta * link.Weight.Value;
         }
 
-        public static int FindMaxValueIndex(IEnumerable<double> xs)
+        public static int FindMaxValueIndex(IEnumerable<float> xs)
         {
             var index = -1;
-            var max = double.MinValue;
+            var max = float.MinValue;
             foreach (var item in xs.Select((x, i) => new { x, i }))
             {
                 if (max < item.x)
